Correção do tipo

+++ database/migrations/2023_02_13_174803_start.php
@@ -445,7 +445,7 @@ return new class extends Migration {
             $table->integer('side');
             $table->integer('strategy');
             $table->integer('player')->constrained("players");
-            $table->integer("battle")->constrained("battles");
+            $table->bigInteger("battle")->constrained("battles");
             $table->foreignId('planet')->constrained("planets");
             $table->string("demage")->nullable();
             $table->bigInteger("start");
